Target and Enmity fixes (#18)
* Fix Reader.Target pointer reads
* Enmity values are uint32

+++ FFXIVAPP.Memory/Reader.PlayerInfo.cs
@@ -55,7 +55,7 @@ namespace FFXIVAPP.Memory
                                 {
                                     ID = (uint) MemoryHandler.Instance.GetPlatformInt(address, MemoryHandler.Instance.Structures.EnmityEntry.ID),
                                     Name = MemoryHandler.Instance.GetString(address + MemoryHandler.Instance.Structures.EnmityEntry.Name),
-                                    Enmity = (uint) MemoryHandler.Instance.GetInt16(address + MemoryHandler.Instance.Structures.EnmityEntry.Enmity)
+                                    Enmity = MemoryHandler.Instance.GetUInt32(address + MemoryHandler.Instance.Structures.EnmityEntry.Enmity)
                                 };
                                 if (enmityEntry.ID > 0)
                                 {
